+++ CHANGELOG.md
@@ -14,8 +14,17 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
-## [29.75.0] - 2025-09-05
-- Add D2 service MethodLevelProperties configuration support
+## [29.75.2] - 2025-09-05
+- Rename the outlier detection field in the D2 cluster
+
+## [29.75.1] - 2025-09-04
+- mark fields as optional
+
+## [29.75.0] - 2025-09-03
+- correct xds client latency metric, add active initial wait time metric, and SD update receipt event
+
+## [29.74.4] - 2025-09-02
+- DataTranslator: fall back to matching simple
 
 ## [29.74.3] - 2025-08-26
 - Add outlier detection config into D2Cluster
@@ -5882,8 +5891,7 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.75.0...master
-[29.75.0]: https://github.com/linkedin/rest.li/compare/v29.74.3...v29.75.0
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.74.3...master
 [29.74.3]: https://github.com/linkedin/rest.li/compare/v29.74.2...v29.74.3
 [29.74.2]: https://github.com/linkedin/rest.li/compare/v29.74.1...v29.74.2
 [29.74.1]: https://github.com/linkedin/rest.li/compare/v29.74.0...v29.74.1